feat: update demographics input
