+++ app-compose/src/main/res/values/themes.xml
@@ -1,7 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.DexCareSampleApp" parent="Theme.AppCompat.DayNight.NoActionBar" >
+    <style name="Theme.DexCareSampleApp" parent="Theme.AppCompat.DayNight.NoActionBar">
         <item name="android:statusBarColor">@color/colorPrimary</item>
+        <item name="android:datePickerDialogTheme">@style/DexCareDatePicker</item>
     </style>
+
+    <style name="DexCareDatePicker" parent="Theme.AppCompat.DayNight.Dialog">
+        <item name="android:datePickerStyle">@style/DatePickerStyle</item>
+    </style>
+
+    <style name="DatePickerStyle" parent="@android:style/Widget.Material.Light.DatePicker">
+        <item name="android:headerBackground">@color/colorPrimary</item>
+    </style>
+
 </resources>
